refactor: introduce typescript compiler
Use TypeScript compiler with the `allowJs` option. With the toolchain
in place we can start migrating the files incrementally.
