+++ src/v2/logging_service_v2_client.js
@@ -20,7 +20,7 @@ const merge = require('lodash.merge');
 const path = require('path');
 const protobuf = require('protobufjs');
 
-const VERSION = require('../../package.json').version;
+const VERSION = require('../../../package.json').version;
 
 /**
  * Service for ingesting and querying logs.
